feat: Improvements to the Fluent Builder to allow the use of the all Where clause methods when writing dynamic SQL.
This commit significantly updates and corrects the documentation across various interfaces and methods within the `Dapper.SimpleSqlBuilder.FluentBuilder` namespace, aiming to improve clarity, accuracy, and usability. Key changes include the correction of reference links, clarification of method combinability, and the use of monospaced font for SQL keywords to enhance readability.

Interface hierarchies have been simplified, notably by changing the base interface of `IDeleteBuilder` and `IJoinBuilder` from `IWhereBuilderEntry` to `IWhereBuilder`, and removing the `IWhereBuilderEntry` and `IWhereFilterBuilderEntry` interfaces. This simplification indicates a consolidation of the interface hierarchy and a focus on streamlining the SQL building process.

Significant changes have also been made to the `FluentSqlBuilder` implementation, including method updates for better clause appending logic and the introduction of a new `entryClause` field to track the initial clause action. The version has been updated to `4.0`, reflecting a major update.

A new set of unit tests has been added to validate the functionality of the `SelectDistinct` method and ensure the correct construction of SQL queries with various clauses. These tests also verify the handling of raw values, formattable strings, custom parameter prefixes, and the enforcement of correct sequence and usage of SQL clauses in the query builder.