Add .NET setup step in codeql.yml
The changes in the `codeql.yml` file include the addition of a new step to set up the .NET environment. This step uses the `actions/setup-dotnet@v4` action and specifies that versions 9.x and 8.x of .NET should be installed. This setup step is added before the initialization of CodeQL tools for scanning.